Refactor and remove test dependency which is not exported
